requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
datasets==3.1.0
dill==0.3.8
docker-pycreds==0.4.0
eval_type_backport==0.2.2
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.1.2
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.1
numpy==1.24.3
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
propcache==0.2.0
protobuf==5.29.3
psutil==7.0.0
pyarrow==17.0.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
sentry-sdk==2.22.0
setproctitle==1.3.4
setuptools==75.1.0
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tokenizers==0.20.3
torch==2.3.0.post101
tqdm==4.66.5
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
wandb==0.19.6
wheel==0.44.0
xxhash==3.5.0
yarl==1.15.2
